Mudando o tipo para melhor aplicação

+++ src/classes/numero_conta.py
@@ -1,5 +1,5 @@
 class Numero_conta:
-    def __init__(self, agencia: int, numero: float) -> None:
+    def __init__(self, agencia: int, numero: int) -> None:
         self.agencia = agencia
         self.numero = numero
